+++ Domibus-MSH/src/main/java/eu/domibus/common/model/security/UserEntityBase.java
@@ -13,7 +13,7 @@ import java.util.Date;
  */
 public interface UserEntityBase extends UserBase {
 
-    int getEntityId();
+    long getEntityId();
 
     UserEntityBase.Type getType();
